restructured the react app
