+++ client/src/components/RecipeListView/index.tsx
@@ -1,0 +1,14 @@
+import React from 'react'
+import RecipeList from './components/RecipeList'
+import AddRecipe from './components/AddRecipe'
+
+const RecipeListView = () => {
+  return (
+    <>
+      <RecipeList />
+      <AddRecipe />
+    </>
+  )
+}
+
+export default RecipeListView
